refactor(mappings): move callback functions to mapping config
Restructures the codebase by moving mapping-related callback functions from
init.lua to config/mappings.lua. This improves code organization and makes
mapping configuration more self-contained.

- Adds callback field to mapping config class
- Moves all mapping callback functions to their respective mapping configs
- Extracts shared functions to utils module
- Simplifies mapping registration by using callback field
